refactored a little, still not working
